% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lab_to_rgb}
\alias{lab_to_rgb}
\title{Convert a cimg object from LAB color space back to RGB}
\usage{
lab_to_rgb(img)
}
\arguments{
\item{img}{A \code{cimg} object in LAB color space (as produced by
\code{\link{rgb_to_lab}} or the internal color-transfer routines).}
}
\value{
A \code{cimg} object in sRGB space, channel values in \[0, 1\].
}
\description{
Inverts \code{\link{rgb_to_lab}}: takes a \code{cimg} whose three color
channels encode CIE L*a*b* values and returns a \code{cimg} in standard
sRGB, with channel values clamped to \[0, 1\].
}
\examples{
\dontrun{
  img_path <- system.file("extdata", "sample_portrait.jpg", package = "CineGrade")
  rgb_img  <- load_image(img_path)
  lab_img  <- rgb_to_lab(rgb_img)
  back     <- lab_to_rgb(lab_img)
}

}
\keyword{internal}
