% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_grade.R
\name{print.CineGrade}
\alias{print.CineGrade}
\title{Print method for CineGrade objects}
\usage{
\method{print}{CineGrade}(x, ...)
}
\arguments{
\item{x}{A \code{CineGrade} object returned by \code{\link{extract_grade}}.}

\item{...}{Further arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for CineGrade objects
}
