% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_grades.R
\name{print.CineGrade_dist}
\alias{print.CineGrade_dist}
\title{Print method for CineGrade_dist objects}
\usage{
\method{print}{CineGrade_dist}(x, ...)
}
\arguments{
\item{x}{A \code{CineGrade_dist} object returned by \code{\link{compare_grades}}.}

\item{...}{Further arguments (unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for CineGrade_dist objects
}
